backend/chatbot-service/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
翻,Complexity:1,Description:
